Move tests
Move the tests which tests functions in pcap.c
to dedicated pcap_tests.c file.
