+++ src/alivedetection_tests.c
@@ -210,213 +210,12 @@ Ensure (alivedetection, get_source_addr_v6)
   close (udpv6soc);
 }
 
-/* If dst for routethrough() is localhost "lo" interface is returned. */
-Ensure (alivedetection, routethrough_dst_is_localhost)
-{
-  /* setup */
-  g_socket_use_real = false;
-  gchar *interface = NULL;
-  gchar *ipv4_str = "127.0.0.1";
-  gvm_host_t *gvm_host = NULL;
-  struct in6_addr dst6;
-  struct in6_addr *dst6_p = &dst6;
-  struct in_addr dst4;
-  struct in_addr *dst4_p = &dst4;
-  assert_that ((gvm_host = gvm_host_from_str (ipv4_str)), is_not_null);
-  assert_that (gvm_host_get_addr6 ((gvm_host_t *) gvm_host, dst6_p),
-               is_equal_to (0));
-  assert_that (dst6_p, is_not_null);
-  dst4.s_addr = dst6_p->s6_addr32[3];
-
-  interface = routethrough (dst4_p, NULL);
-  (void) interface;
-
-  /* dependent on local environment */
-  // assert_that ((interface = routethrough (dst4_p, NULL)), is_not_null);
-  // assert_that (interface, is_equal_to_string ("lo"));
-  g_socket_use_real = true;
-}
-
-/* If dst is not null for routethrough() then another interface than "lo" is
- * returned. */
-Ensure (alivedetection, routethrough_dst_is_not_localhost)
-{
-  g_socket_use_real = false;
-  /* setup */
-  gchar *interface = NULL;
-  gchar *ipv4_str = "93.184.216.34"; /* example.com */
-  gvm_host_t *gvm_host = NULL;
-  struct in6_addr dst6;
-  struct in6_addr *dst6_p = &dst6;
-  struct in_addr dst4;
-  struct in_addr *dst4_p = &dst4;
-  assert_that ((gvm_host = gvm_host_from_str (ipv4_str)), is_not_null);
-  assert_that (gvm_host_get_addr6 ((gvm_host_t *) gvm_host, dst6_p),
-               is_equal_to (0));
-  assert_that (dst6_p, is_not_null);
-  dst4.s_addr = dst6_p->s6_addr32[3];
-
-  interface = routethrough (dst4_p, NULL);
-  assert_that (interface, is_not_equal_to_string ("lo"));
-  g_socket_use_real = true;
-}
-
-/* If neither dst nor src address are given to routethrough NULL is returned. */
-Ensure (alivedetection, routethrough_no_src_dst_given)
-{
-  gchar *interface = NULL;
-  assert_that ((interface = routethrough (NULL, NULL)), is_null);
-}
-
-/* If global_source_addr is present then routethrough writes it into src. */
-Ensure (alivedetection, routethrough_src_globalsource_set)
-{
-  /* setup */
-  g_socket_use_real = false;
-  cgreen_mocks_are (learning_mocks);
-
-  struct in_addr src = {.s_addr = 0}; /* ip src */
-  gchar *interface = NULL;
-  struct in_addr dst;
-  inet_pton (AF_INET, "93.184.216.34", &(dst.s_addr));
-
-  /* global source address set */
-  gvm_source_iface_init ("lo"); // lo is set but not really used after being set
-  /* dst not given */
-  assert_that ((interface = routethrough (NULL, &src)), is_null);
-  assert_that ((src.s_addr == INADDR_ANY));
-  /* dst localhost given */
-  src.s_addr = 0;
-
-  interface = routethrough (&dst, &src);
-  /* dependent on local environment */
-  // assert_that ((interface = routethrough (&dst, &src)), is_not_null);
-  assert_that (interface, is_not_equal_to_string ("lo"));
-  assert_that ((src.s_addr != INADDR_ANY));
-  g_socket_use_real = true;
-}
-
-/* If global_source_addr is not present then routethrough writes it into src. */
-Ensure (alivedetection, routethrough_src_globalsource_not_set)
-{
-  g_socket_use_real = false;
-
-  struct in_addr src = {.s_addr = 0}; /* ip src */
-  gchar *interface = NULL;
-  struct in_addr dst;
-  inet_pton (AF_INET, "127.0.0.1", &(dst.s_addr));
-
-  /* global source address not set */
-  gvm_source_iface_init (NULL);
-  /* dst not given */
-  assert_that ((interface = routethrough (NULL, &src)), is_null);
-  assert_that ((src.s_addr == INADDR_ANY));
-  /* dst localhost given */
-  src.s_addr = 0;
-
-  interface = routethrough (&dst, &src);
-  /* dependent on local environment */
-  // assert_that ((interface = routethrough (&dst, &src)), is_not_null);
-  // assert_that (interface, is_equal_to_string ("lo"));
-  assert_that ((src.s_addr != INADDR_ANY));
-  g_socket_use_real = true;
-}
-
-Ensure (alivedetection, gvm_source_addr)
-{
-  struct in_addr src;
-
-  /* global source address not set */
-  gvm_source_iface_init (NULL);
-  gvm_source_addr (&src);
-  assert_that ((src.s_addr == INADDR_ANY));
-
-  /* global source address */
-  gvm_source_iface_init ("lo");
-  gvm_source_addr (&src);
-  assert_that ((src.s_addr != INADDR_ANY));
-}
-
-Ensure (alivedetection, v6_islocalhost)
-{
-  /* IPv4 */
-  struct in_addr addr;
-  struct sockaddr_in sin;
-  memset (&sin, 0, sizeof (struct sockaddr_in));
-  sin.sin_family = AF_INET;
-
-  /* example.com */
-  inet_pton (AF_INET, "93.184.216.34", &(addr.s_addr));
-
-  /* IPv6 */
-  struct in6_addr addr_6;
-
-  inet_pton (AF_INET6, "::FFFF:127.0.0.1", &(addr_6));
-  assert_that (v6_islocalhost (&addr_6), is_true);
-  inet_pton (AF_INET6, "::FFFF:0.0.0.0", &(addr_6));
-  assert_that (v6_islocalhost (&addr_6), is_true);
-  inet_pton (AF_INET6, "::FFFF:127.100.5.99", &(addr_6));
-  assert_that (v6_islocalhost (&addr_6), is_true);
-  /* loopback address */
-  inet_pton (AF_INET6, "0:0:0:0:0:0:0:1", &(addr_6));
-  assert_that (v6_islocalhost (&addr_6), is_true);
-
-  /* dependent on local environment */
-  // inet_pton (AF_INET6, <some local interface address>, &(addr_6));
-  // assert_that (v6_islocalhost (&addr_6), is_true);
-
-  /* example.com */
-  inet_pton (AF_INET6, "2606:2800:220:1:248:1893:25c8:1946", &(addr_6));
-  assert_that (v6_islocalhost (&addr_6), is_false);
-}
-
-Ensure (alivedetection, islocalhost)
-{
-  /* IPv4 */
-  struct in_addr addr;
-
-  inet_pton (AF_INET, "127.0.0.1", &(addr.s_addr));
-  assert_that (islocalhost (&addr), is_true);
-  inet_pton (AF_INET, "0.0.0.0", &(addr.s_addr));
-  assert_that (islocalhost (&addr), is_true);
-  inet_pton (AF_INET, "127.100.5.99", &(addr.s_addr));
-  assert_that (islocalhost (&addr), is_true);
-
-  /* dependent on local environment */
-  // // inet_pton (AF_INET, <some local interface address>, &(addr));
-  // // assert_that (islocalhost (&addr), is_true);
-
-  /* example.com */
-  inet_pton (AF_INET, "93.184.216.34", &(addr.s_addr));
-  assert_that (islocalhost (&addr), is_false);
-}
-
-TestSuite *
-openvas_routethrough ()
-{
-  TestSuite *suite = create_test_suite ();
-  add_test_with_context (suite, alivedetection, gvm_source_addr);
-  add_test_with_context (suite, alivedetection, routethrough_dst_is_localhost);
-  add_test_with_context (suite, alivedetection,
-                         routethrough_dst_is_not_localhost);
-  add_test_with_context (suite, alivedetection, routethrough_no_src_dst_given);
-  add_test_with_context (suite, alivedetection,
-                         routethrough_src_globalsource_set);
-  add_test_with_context (suite, alivedetection,
-                         routethrough_src_globalsource_not_set);
-  add_test_with_context (suite, alivedetection, v6_islocalhost);
-  add_test_with_context (suite, alivedetection, islocalhost);
-
-  return suite;
-}
-
 int
 main (int argc, char **argv)
 {
   TestSuite *suite;
 
   suite = create_test_suite ();
-  add_suite (suite, openvas_routethrough ());
 
   add_test_with_context (suite, alivedetection, fill_ports_array);
   add_test_with_context (suite, alivedetection, set_all_needed_sockets);
